refactor(core): reduce invoker duplication
