+++ wow-core/src/main/java/me/ahoo/wow/infra/invoker/MethodHandleStaticFunctionInvoker.java
@@ -16,176 +16,17 @@ package me.ahoo.wow.infra.invoker;
 import java.lang.invoke.MethodHandle;
 import java.lang.reflect.Method;
 
-final class MethodHandleStaticFunctionInvoker implements StaticFunctionInvoker {
+final class MethodHandleStaticFunctionInvoker extends AbstractMethodHandleReceiverlessFunctionInvoker
+        implements StaticFunctionInvoker {
     private final Method method;
-    private final MethodHandle handle;
-    private final int parameterCount;
 
     MethodHandleStaticFunctionInvoker(Method method, MethodHandle handle) {
+        super(handle, method.getParameterCount());
         this.method = method;
-        this.handle = handle;
-        this.parameterCount = method.getParameterCount();
     }
 
     @Override
-    public int parameterCount() {
-        return parameterCount;
-    }
-
-    @Override
-    public Object invoke(Object[] args) throws Throwable {
-        Object[] actualArgs = FunctionInvocationSupport.actualArgs(args);
-        try {
-            if (actualArgs.length <= 9) {
-                return FunctionInvocationSupport.invokeByArgumentArray(this, actualArgs);
-            }
-            return handle.invokeWithArguments(actualArgs);
-        } catch (Throwable error) {
-            throw normalize(error, actualArgs);
-        }
-    }
-
-    @Override
-    public Object invoke0() throws Throwable {
-        try {
-            return handle.invoke();
-        } catch (Throwable error) {
-            throw normalize(error, FunctionInvocationSupport.EMPTY_ARGS);
-        }
-    }
-
-    @Override
-    public Object invoke1(Object arg1) throws Throwable {
-        try {
-            return handle.invoke(arg1);
-        } catch (Throwable error) {
-            throw normalize(error, arg1);
-        }
-    }
-
-    @Override
-    public Object invoke2(Object arg1, Object arg2) throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2);
-        }
-    }
-
-    @Override
-    public Object invoke3(Object arg1, Object arg2, Object arg3) throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2, arg3);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2, arg3);
-        }
-    }
-
-    @Override
-    public Object invoke4(Object arg1, Object arg2, Object arg3, Object arg4) throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2, arg3, arg4);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2, arg3, arg4);
-        }
-    }
-
-    @Override
-    public Object invoke5(Object arg1, Object arg2, Object arg3, Object arg4, Object arg5) throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2, arg3, arg4, arg5);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2, arg3, arg4, arg5);
-        }
-    }
-
-    @Override
-    public Object invoke6(Object arg1, Object arg2, Object arg3, Object arg4, Object arg5, Object arg6)
-            throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2, arg3, arg4, arg5, arg6);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2, arg3, arg4, arg5, arg6);
-        }
-    }
-
-    @Override
-    public Object invoke7(Object arg1, Object arg2, Object arg3, Object arg4, Object arg5, Object arg6,
-                          Object arg7) throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2, arg3, arg4, arg5, arg6, arg7);
-        }
-    }
-
-    @Override
-    public Object invoke8(Object arg1, Object arg2, Object arg3, Object arg4, Object arg5, Object arg6,
-                          Object arg7, Object arg8) throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8);
-        }
-    }
-
-    @Override
-    public Object invoke9(Object arg1, Object arg2, Object arg3, Object arg4, Object arg5, Object arg6,
-                          Object arg7, Object arg8, Object arg9) throws Throwable {
-        try {
-            return handle.invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9);
-        } catch (Throwable error) {
-            throw normalize(error, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9);
-        }
-    }
-
-    private Throwable normalize(Throwable error, Object[] args) {
+    protected Throwable normalize(Throwable error, Object... args) {
         return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, args);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2, Object arg3) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2, arg3);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2, Object arg3, Object arg4) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2, arg3,
-            arg4);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2, Object arg3, Object arg4, Object arg5) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2, arg3,
-            arg4, arg5);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2, Object arg3, Object arg4, Object arg5,
-                                Object arg6) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2, arg3,
-            arg4, arg5, arg6);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2, Object arg3, Object arg4, Object arg5,
-                                Object arg6, Object arg7) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2, arg3,
-            arg4, arg5, arg6, arg7);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2, Object arg3, Object arg4, Object arg5,
-                                Object arg6, Object arg7, Object arg8) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2, arg3,
-            arg4, arg5, arg6, arg7, arg8);
-    }
-
-    private Throwable normalize(Throwable error, Object arg1, Object arg2, Object arg3, Object arg4, Object arg5,
-                                Object arg6, Object arg7, Object arg8, Object arg9) {
-        return FunctionInvocationSupport.normalizeInvocationException(error, method, true, null, arg1, arg2, arg3,
-            arg4, arg5, arg6, arg7, arg8, arg9);
     }
 }
